created a new activity with theme selection
