+++ .idea/misc.xml
@@ -12,10 +12,11 @@
         <entry key="..\:/Users/torte/AndroidStudioProjects/SuperCalculate/app/src/main/res/drawable/gradient_list.xml" value="0.1505" />
         <entry key="..\:/Users/torte/AndroidStudioProjects/SuperCalculate/app/src/main/res/drawable/gradient_list2.xml" value="0.1505" />
         <entry key="..\:/Users/torte/AndroidStudioProjects/SuperCalculate/app/src/main/res/layout/activity_main.xml" value="0.264" />
+        <entry key="..\:/Users/torte/AndroidStudioProjects/SuperCalculate/app/src/main/res/layout/activity_setting.xml" value="0.4" />
       </map>
     </option>
   </component>
-  <component name="ProjectRootManager" version="2" languageLevel="JDK_11" default="true" project-jdk-name="Android Studio default JDK" project-jdk-type="JavaSDK">
+  <component name="ProjectRootManager" version="2" languageLevel="JDK_X" project-jdk-name="Android Studio default JDK" project-jdk-type="JavaSDK">
     <output url="file://$PROJECT_DIR$/build/classes" />
   </component>
   <component name="ProjectType">
