+++ app/src/main/java/ru/gp/supercalculate/MainActivity.java
@@ -3,24 +3,36 @@ package ru.gp.supercalculate;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.constraintlayout.widget.ConstraintLayout;
 
+import android.content.Intent;
 import android.content.SharedPreferences;
 import android.graphics.drawable.AnimationDrawable;
 import android.os.Bundle;
 import android.view.View;
+import android.widget.Button;
 import android.widget.RadioGroup;
 
+import com.google.android.material.button.MaterialButton;
 import com.google.android.material.radiobutton.MaterialRadioButton;
 
 public class MainActivity extends BaseActivity {
-
+    MaterialButton settButton;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
+        settButton = findViewById(R.id.settingButton);
+        settButton.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                Intent intent = new Intent(MainActivity.this, SettingActivity.class);
+                startActivity(intent);
+                recreate();
+
+            }
+        });
         Animation1();
-        initThemeChooser();
 
     }
 
@@ -30,16 +42,10 @@ public class MainActivity extends BaseActivity {
         animationDrawable.setEnterFadeDuration(2500);
         animationDrawable.setExitFadeDuration(5000);
         animationDrawable.start();
+
+
     }
-    // Инициализация радиокнопок
-    private void initThemeChooser() {
-        initRadioButton(findViewById(R.id.standard),
-                MyStandard);
-        initRadioButton(findViewById(R.id.ocean),
-                MyOcean);
-        RadioGroup rg = findViewById(R.id.radioButtons);
-        ((MaterialRadioButton)rg.getChildAt(getCodeStyle(MyStandard))).setChecked(true);
-    }
+
 // Все инициализации кнопок очень похожи, поэтому создадим метод для переиспользования
     private void initRadioButton(View button, final int codeStyle) {
         button.setOnClickListener(new View.OnClickListener() {
