+++ app/src/main/java/ru/gp/supercalculate/SettingActivity.java
@@ -1,0 +1,50 @@
+package ru.gp.supercalculate;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.os.Bundle;
+import android.view.View;
+import android.widget.RadioGroup;
+
+import com.google.android.material.button.MaterialButton;
+import com.google.android.material.radiobutton.MaterialRadioButton;
+
+public class SettingActivity extends BaseActivity {
+
+    MaterialButton bckBtm;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_setting);
+
+        initThemeChooser();
+
+        bckBtm = findViewById(R.id.backBtm);
+        bckBtm.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                finish();
+            }
+        });
+    }
+    private void initThemeChooser() {
+        initRadioButton(findViewById(R.id.standard),
+                MyStandard);
+        initRadioButton(findViewById(R.id.ocean),
+                MyOcean);
+        RadioGroup rg = findViewById(R.id.radioButtons);
+        ((MaterialRadioButton) rg.getChildAt(getCodeStyle(MyStandard))).setChecked(true);
+    }
+    private void initRadioButton(View button, final int codeStyle) {
+        button.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+// сохраним настройки
+                setAppTheme(codeStyle);
+// пересоздадим активити, чтобы тема применилась
+                recreate();
+            }
+        });
+    }
+}
