+++ app/src/main/res/values/strings.xml
@@ -1,6 +1,8 @@
 <resources>
     <string name="app_name">SuperCalculate</string>
 
+    <string name="displayHint">0</string>
+
     <string name="one">1</string>
     <string name="two">2</string>
     <string name="three">3</string>
@@ -24,8 +26,8 @@
 
     <string name="standard">Standard Theme</string>
     <string name="ocean">Ocean Theme</string>
-
-
+    <string name="theme">Theme</string>
+    <string name="back">Back</string>
 
 
 </resources>
